requirements.txt
gitdb~=4.0.11
requests~=2.31.0